requirements.txt
asgiref==3.7.2
Django==5.0
django-multiupload==0.6.1
Pillow==10.1.0
sqlparse==0.4.4
typing_extensions==4.8.0
